% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{BetaGamma}
\alias{BetaGamma}
\title{Generic selector using the Beta-Gamma space by Brown et al.}
\usage{
BetaGamma(X, Y, k, beta, gamma)
}
\arguments{
\item{X}{Predictor table, given as a matrix or data frame with only factor or logical columns. \code{NA}s are not allowed.}

\item{Y}{Decision attribute; must be either a factor or a logical vector. \code{NA}s are not allowed.}

\item{k}{Number of attributes to select; first element will be used. Must not exceed \code{ncol(X)}.}

\item{beta}{Beta parameter; see the reference for details.}

\item{gamma}{Gamma parameter; see the reference for details.}
}
\value{
A list with two elements: \code{selection}, a vector of names of the selected features (note that it may be shorter than \code{k}), and \code{scores}, a vector of feature scores.
}
\description{
Generic selector using the Beta-Gamma space by Brown et al.
}
\examples{
BetaGamma(data.frame(lapply(iris[,-5],cut,10)),iris[,5],4,0.1,0.1)
}
\references{
"Conditional Likelihood Maximisation: A Unifying Framework for Mutual Information Feature Selection" G. Brown, A. Pocock, M.-J. Zhao, M. Lujan, JMLR (2011)
}
