% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{CondMI}
\alias{CondMI}
\title{CMI selection using a greedy forward search.}
\usage{
CondMI(X, Y, k)
}
\arguments{
\item{X}{Predictor table, given as a matrix or data frame with only factor or logical columns. \code{NA}s are not allowed.}

\item{Y}{Decision attribute; must be either a factor or a logical vector. \code{NA}s are not allowed.}

\item{k}{Number of attributes to select; first element will be used. Must not exceed \code{ncol(X)}.}
}
\value{
A list with two elements: \code{selection}, a vector of names of the selected features (note that it may be shorter than \code{k}), and \code{scores}, a vector of feature scores.
}
\description{
CMI selection using a greedy forward search.
}
\examples{
CondMI(data.frame(lapply(iris[,-5],cut,10)),iris[,5],4)
}
