% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{DISR}
\alias{DISR}
\title{Double Input Symmetrical Relevance selection.}
\usage{
DISR(X, Y, k)
}
\arguments{
\item{X}{Predictor table, given as a matrix or data frame with only factor or logical columns. \code{NA}s are not allowed.}

\item{Y}{Decision attribute; must be either a factor or a logical vector. \code{NA}s are not allowed.}

\item{k}{Number of attributes to select; first element will be used. Must not exceed \code{ncol(X)}.}
}
\value{
A vector of names of the selected features; note that it may be shorter than \code{k}, but should always contain at least one feature.
}
\description{
Double Input Symmetrical Relevance selection.
}
\examples{
DISR(data.frame(lapply(iris[,-5],cut,10)),iris[,5],4)
}
\references{
"On the Use of Variable Complementarity for Feature Selection in Cancer Classification" P. Meyer and G. Bontempi, (2006)
}

