% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{JMI}
\alias{JMI}
\title{Joint Mutual Information selection.}
\usage{
JMI(X, Y, k)
}
\arguments{
\item{X}{Predictor table, given as a matrix or data frame with only factor or logical columns. \code{NA}s are not allowed.}

\item{Y}{Decision attribute; must be either a factor or a logical vector. \code{NA}s are not allowed.}

\item{k}{Number of attributes to select; first element will be used. Must not exceed \code{ncol(X)}.}
}
\value{
A list of names of the selected features.
}
\description{
Joint Mutual Information selection.
}
\examples{
JMI(data.frame(lapply(iris[,-5],cut,10)),iris[,5],4)
}
\references{
"Data Visualization and Feature Selection: New Algorithms for Nongaussian Data H. Yang and J. Moody, NIPS (1999)
}

