% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{mRMR_D}
\alias{mRMR_D}
\title{Minimum Relevance Maximum Redundancy selection using the difference variant by H. Peng et al.}
\usage{
mRMR_D(X, Y, k)
}
\arguments{
\item{X}{Predictor table, given as a matrix or data frame with only factor or logical columns. \code{NA}s are not allowed.}

\item{Y}{Decision attribute; must be either a factor or a logical vector. \code{NA}s are not allowed.}

\item{k}{Number of attributes to select; first element will be used. Must not exceed \code{ncol(X)}.}
}
\value{
A list of names of the selected features.
}
\description{
Minimum Relevance Maximum Redundancy selection using the difference variant by H. Peng et al.
}
\examples{
mRMR_D(data.frame(lapply(iris[,-5],cut,10)),iris[,5],4)
}
\references{
"Feature Selection Based on Mutual Information: Criteria of Max-Dependency, Max-Relevance, and Min-Redundancy" H. Peng et al. IEEE Pattern Analysis and Machine Intelligence (PAMI) (2005)
}

